feat: create useEndpointsSetting.ts
